requirements/base.txt
asgi-redis==1.4.3
channels==1.1.8
daphne==1.3.0
Django==1.11.5
django-ckeditor==5.3.0
django-registration==2.2
django-registration-redux==1.7
gunicorn==19.7.1
ipdb==0.10.3
ipython==6.2.0
mysqlclient==1.3.12
Pillow==4.2.1
psycopg2==2.7.3.1
python-dotenv==0.7.1
redis==2.10.6
Twisted[tls,http2]
